requirements.txt
# 核心依赖
gradio>=3.50.0
pdfminer.six>=20221105
PyMuPDF>=1.24.0
sentence-transformers>=2.2.2
faiss-cpu
langchain-text-splitters
python-dotenv>=1.0.0
requests>=2.31.0
markdown>=3.4.3
fastapi>=0.104.1
uvicorn>=0.23.2
python-multipart>=0.0.6
urllib3>=1.26.16
jieba>=0.42.1
pydantic>=2.9.2,<2.11.0

# 处理相关的pdf、word等文件所需相关库
python-docx
python-pptx

# 工具依赖
numpy>=1.26.1
rank-bm25>=0.2.2

# 系统监控
psutil>=5.9.0
